feat: исправлено поиск поппулярных на значение райтинга + немного исправлений в схеме и методах

+++ src/main/java/ru/yandex/practicum/filmorate/model/Rate.java
@@ -21,5 +21,5 @@ public class Rate {
     /**
      * Поле содержащее оценку фильма от пользователя
      */
-    private Double rate;
+    private int rate;
 }
